Revert "Merge pull request #179 from Financial-Times/matth/remove-old-welcome"
This reverts commit d70ec5d0efbefc2ed7ae0c2d4695805632bd58bf, reversing
changes made to a889837fe36631aac020847ece54c95bd4d347dd.
