+++ main.js
@@ -102,8 +102,11 @@ export function bootstrap (cb) {
 					cookieMessage.init();
 				}
 
-				if (opts.welcomeMessage && flags.get('welcomePanel')) {
-					welcomeMessage.init();
+				if (opts.welcomeMessage) {
+					let version = flags.get('newFooter') ? 'new' : 'old';
+					flags.get('welcomePanel') && welcomeMessage[version].init({
+						enableOverlay: flags.get('myFTOnboardingOverlay')
+					});
 				}
 
 				if (opts.messagePrompts) {
