+++ welcome-message/new/_message.html
@@ -1,7 +1,7 @@
 <div class="n-welcome-message__wrapper">
 	<div class="n-welcome-message__column n-welcome-message__column--wrapper">
 		<h6 class="n-welcome-message__heading">
-			Welcome to the new<span class="n-welcome-message__extended">, improved</span> FT website
+			Welcome to the new FT.com
 		</h6>
 		<p class="n-welcome-message__intro">
 			We're constantly improving and would like to hear from you.
@@ -9,11 +9,11 @@
 	</div>
 	<div class="n-welcome-message__column n-welcome-message__column--actions">
 		<a class="n-welcome-message__button n-welcome-message__button--feedback" data-tracking="nps-feedback" href="/nps-feedback">
-			<span class="n-welcome-message__extended">Provide</span> Feedback
+			<span class="n-welcome-message__button-extended">Provide</span> Feedback
 		</a>
 		{{#unless @root.forceOptInDevice}}
 		<a class="n-welcome-message__button n-welcome-message__button--opt-out js-optout-link" data-tracking="opt-out" href="/opt-out-confirm">
-			Return to <span class="n-welcome-message__extended">the old</span> FT.com
+			Return to <span class="n-welcome-message__button-extended">the old</span> FT.com
 		</a>
 		{{/unless}}
 		<a class="n-welcome-message__button n-welcome-message__button--webapp is-hidden js-webapp-link" data-tracking="visit-webapp" href="https://app.ft.com">
