+++ welcome-message/old/main.scss
@@ -1,0 +1,126 @@
+@if $_n-ui-welcome-message-is-silent == false {
+	.n-welcome {
+		@include oTypographySansSize(s);
+		font-family: MetricWeb, Arial, sans-serif;
+		background: getColor('grey-tint5');
+		color: getColor('pink');
+		padding: 10px 0 5px;
+		border-top: 5px solid getColor('grey-tint3');
+		position: relative;
+		display: block;
+
+		@include oGridRespondTo(L) {
+			padding: 15px 0;
+
+			.o-grid-container {
+				& > div {
+					display: flex;
+					align-items: center;
+				}
+			}
+		}
+
+		&__heading {
+			@include oTypographySansSize(m);
+			font-style: normal;
+			font-weight: 500;
+			line-height: 1;
+			margin: 0;
+
+			@include oGridRespondTo(M) {
+				font-size: 23px;
+			}
+		}
+
+		&__link {
+			border-bottom: 0;
+			color: getColor('pink');
+			text-decoration: underline;
+			line-height: 1;
+			padding: 10px 0 3px;
+
+			@include oGridRespondTo(M) {
+				padding: 5px 0 0;
+			}
+		}
+
+		&__intro {
+			@include oTypographySans(xs);
+			margin: 0.3em 0;
+			max-width: 90%;
+		}
+
+		&__intro &__link {
+			@include oTypographySansSize(xs);
+		}
+
+		&--fixed {
+			width: 100%;
+			position: fixed;
+			bottom: 0;
+			display: none;
+			visibility: hidden;
+		}
+
+		&--fixed-visible {
+			@include nUtilZIndexFor('banner');
+			display: block;
+			visibility: visible;
+		}
+
+		&--hidden {
+			visibility: hidden;
+		}
+	}
+
+	.n-welcome__btns {
+		margin: 5px 0 10px;
+		list-style: none;
+		overflow: hidden;
+
+		@include oGridRespondTo(L) {
+			margin: 0;
+		}
+
+		& > li {
+			float: left;
+			margin-top: 5px;
+			margin-right: 5px;
+
+			@include oGridRespondTo(L) {
+				margin-top: 0;
+				margin-right: 10px;
+			}
+
+			&:last-child {
+				margin-right: 0;
+			}
+		}
+	}
+
+	.n-welcome__btn {
+		@include nButtons($theme: inverse);
+
+		@include oGridRespondTo(M) {
+			@include nButtonsSize(big);
+		}
+
+		&--standout {
+			@include nButtons($theme: standout);
+		}
+	}
+
+	.n-welcome__close {
+		display: none;
+
+		.n-welcome--fixed & {
+			display: inline-block;
+		}
+	}
+
+	@media print {
+		.n-welcome {
+			display: none;
+		}
+	}
+}
